pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "imgflip-meme-mcp"
version = "0.1.0"
description = "MCP server for intelligent meme generation using Imgflip API"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "mcp>=1.0.0",
    "requests>=2.31.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
]
requires-python = ">=3.8"

[project.scripts]
imgflip-mcp = "imgflip_meme_mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/imgflip_meme_mcp"]

[tool.hatch.envs.default]
dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.ruff]
select = ["E", "F", "I", "N", "W"]
ignore = []
line-length = 88
target-version = "py38"